Use a modified version of the cases tab
This decorator now utilises the cases tab within the tab builder but
with some distinct differences. There is a method to strip away the
create option in the dropdown which isn't needed for this part of the
site. It also changes the text of the dropdown to indicate it is for all
cases.