Import data after deploy on staging
Rather than before. This ensures that the full deploy has already
successfully occurred and so any needed gems for the production import
are available within the container, which means things will be less
likely to blow up when attempting the import due to missing gems (which
just happened to me).