Vary styling of ChangeMotdRequest button
Based on whether the ChangeMotdRequest has already been applied - if it
has it probably doesn't need to be re-applied, and though this should
work in this case, for future similar types of requests that we plan to
add (e.g. user deletions) it may not. Therefore for both these reasons
visually indicate this as a riskier action in this case.