Make type heirarchy of `MaintenanceWindow` consistent.
There are two places in the code where we define a `MaintenanceWindow`
class: the model itself, and in this validator, where we open up the
class and insert the `Validator` class. However, in the latter case we
weren't specifying a superclass for `MaintenanceWindow`, which meant
that `nullalign` was getting very upset about an inconsistent type
hierarchy.