Make obfuscating real contact emails work again
We recently attempted to change how we anonymized real contact emails in
staging so that the email for the contact with name 'Some Contact' and
email `some.contact.email@example.com` became
`center+somecontact@alces-software.com` rather than
`some.contact.email@alces-software.com`, as was done previously.

This was done by taking the user's name and removing any spaces within
this, and then using `center+` + this as the local part for their new
email.

However, when run on our production data to convert this for use in
staging, this failed as many characters can appear in our User names
which would not be valid in the local part of an email address, e.g. `(`
and `)`, and so we were attempting to save various Users with invalid
email addresses.

Since the only thing we really care about here is having email addresses
from staging in the format `center+...@alces-software.com`, so these are
easily filterable from real support messages, we can just revert to
using the original local part when forming modified email addresses,
just with the `center+` now added.