Correct method call syntax for Ruby upgrade
Ruby 2.4.2+ makes calling a method with arguments, without parentheses,
and with a curly brace block, a syntax error. See
https://bugs.ruby-lang.org/issues/13939,
https://bugs.ruby-lang.org/issues/13976,
https://bugs.ruby-lang.org/issues/13898, etc. for some issues related to
this.

We do this in many places, so need to change all these places to fix
this so our code will run when we upgrade Ruby past this version. I
think this looks slightly worse, but what do I know, and I can't argue
with the Ruby interpreter anyway.