Remove loading `active_storage`
After updating to Rails 5.2, which now includes `active_storage`, our
rails-admin interface started erroring, as it was attempting to load
`active_storage`, however `active_storage` requires a new table be
created for the stored objects (which we did not yet have).

Since we're not actually using `active_storage` at the moment, rather
than loading it and creating an empty table for it to use lets just
disable it. If we do later need it we can just add it back in.