Refactor the method for returning component types
As the component_group_for_types method got extracted and refactored for
its main purpose the only place it was being used was for the
'Components' dropdown menu that displayed the available types.

Because of this the method has been renamed and refactored to only
return the unique component group types for displaying on the dropdown.

Changes have been made to the associated decorator and test spec to use
the new 'available_component_group_types' method.