Improve Case form UI following tabbed Tiers
Move things around a bit, and remove now unnecessary divider, to improve
both the code and Case form UI now that Tiers are shown in tabs.

Note that the chargeable Tier alert looks slightly bad at the moment due
to a bug in one of our dependencies
(https://github.com/rundis/elm-bootstrap/issues/105), but should look
slightly better if/when my fix for this is merged and we upgrade to a
version with this in.