Dynamically create paths for cases in links
This should allow for the case links to point the user to the page for
the related case within the same dashboard. Therefore if you were on the
cluster dashboard and went to view a case it wouldn't take you away from
the context of a cluster.

Hopefully this should make things less confusing as the user isn't being
jumped around to different dashboards unexpectedly.