Order Cluster component_groups by type ordering
Rather than needing to respecify that they should be in the defined
order each time, just scope the ordering to the relation so accessing
the `component_groups` for a Cluster always gives them in the order we
want.

Actually doing this was slightly trickier than expected as the Cluster's
`components`, accessed through the `component_groups`, also pick up the
same ordering, which just caused an error as this makes no sense in this
case; to fix this we also need to `unscope` when getting the
`components`.

As suggested in
https://github.com/alces-software/alces-flight-center/pull/302#discussion_r190673717.