Modify User data after running migrations
When modifying imported production data for use in development, modify
the imported Users after rather than before we run all Rails and data
migrations. This way the current database schema and data should be
consistent with the code, and so there is less possibility of User
validations failing when trying to update Users (which just happened to
me).