Move email_recipients to Site; replace all_contacts
Move `Case#email_recipients` to Site, and then just delegate to this
from Case, which removes need to have this logic unnecessarily split
across 2 classes. Doing this also means can get rid of the
`Site#all_contacts` method (which is quite badly named since the
addition of Users with a `viewer` `role`, which will also be included).