Extract polymorphic_path logic to decorators
This commit is going to look messy as there were a bunch of changes to
many files that now use the new method `dashboard_case_path` to handle
redirecting to the correct case page for the current context. Or the
cluster page for that case if the user is at the top level.

Amongst these changes are some other small changes that should hopefully
be fairly self-explanatory.