Use more correct method
Using `readable_model_name` worked in these two cases with the current
models they were using, but would stop working as soon as we use a model
with more than one 'hump' in its CamelCased class name (e.g.
ComponentGroup) as this would not correctly be snake_cased, which is
what we need in these situations.