Have (most) cron jobs run in reused container
Re-use single `cron` container for most Flight Center cron jobs, to
avoid large resource use of starting many containers.

The backup job will remain as it for now since it has been working fine
and want to avoid potentially breaking it, and in any case it only
launches one container every three hours, which shouldn't be a major
issue.

This also seems consistent with the Dokku docs for when a task should
launch vs reuse a container: "For tasks that should not be interrupted,
run is the preferred method of handling cron tasks, as the container
will continue running even during a deploy or scaling event.  The
trade-off is that there will be an increase in memory usage if there are
multiple concurrent tasks running." (from
http://dokku.viewdocs.io/dokku/deployment/one-off-processes/#using-run-for-cron-tasks)

As suggested at
https://alces.slack.com/archives/C0BDJJXCL/p1527755652000350, and
following instructions and reasoning at
http://dokku.viewdocs.io/dokku/deployment/one-off-processes/#entering-existing-containers.