Handle Slack::Notifier pings in a method
This changes all notifier pings into method calls to the new 'send'
method that takes the note, used as the message attachment, as an
argument and handles the ping in one location.

In future this will allow for easier manipulation of all messages if
necessary e.g. setting the channel pings are sent to for testing/dev
purposes so that the support channel isn't spammed by us pesky
developers.