Merge SSO docs into rest of README
So README includes all steps needed in each section to setup/work with a
Flight Center development environment. Some parts of the original SSO
README have been removed/reworded to fit in with the rest of the README,
or where things are already covered in this or the `flight-sso` README.

Documenting the SSO environment variables has also been moved to the
`.env.example` file, to keep all Flight Center environment variables
documented in one place.