Clean notifications and extract logic to methods
This commit cleans all Slack notifications made by this notifier. Tweaks
can probably be made to clean them up further but for now they will send
the relevant information with links to the appropriate pages.

Also I have moved some repeat logic to private methods to avoid
repetition and make it hopefully easier to adjust the notifications if
and when is necessary.