Remove weird asymmetry between Case index actions
These actions are doing basically the same thing, but looked quite
different before, and it is also quite unusual to have a controller
action take arguments which are only needed so it can be called by
another action. Having them both call a third method instead makes this
clearer and less weird.