Extract class to encapsulate Case commenting
Encapsulates whether Case can be commented on, as well as the reason
why. Possibly slightly overkill to extract this at the moment, but I
need a common place which can be used both in the view and from the
soon-to-be-added CaseCommentPolicy (where there is also no access to
`current_user`); I also expect this behaviour to soon be expanded to
handle when and how commenting should be disabled for viewers.