Add a link when restricting text length
Whenever text is longer than 80 characters only the first 80 are
displayed followed by '...' to indicate there are more details than is
being displayed.

The '...' will now be clickable and link to the relevant location for
the notification e.g. the case page.

Another minor change here is that some pretext has been removed as it
was essentially repeating what the notification attachment stated. This
could be added back easily if things seem unclear however.