Extract case path url methods from decorators
Previously having these methods within their respective decorators
caused issues in some cases. This was due to the way the context is
handled within Draper, meaning the helpers it was attached to were
changing as the application moved between contexts.

Having this logic handled within the models means that we don't need to
apply band-aids as and when we find issues relating to this.