Implement CaseCommentPolicy and use in controller
This is currently a thin wrapper around `CaseCommenting#disabled?`,
since the same logic is needed both when authorizing CaseComment
creation and elsewhere.

Also added a method to Case to check whether a User can comment, to be
used from the CaseCommentPolicy, which made testing this all works
simpler and makes sense that this should live here.