Assign Slack environment variables
There are three environment variables relating to the Slack integration
that are important to be aware of; webhook_url, channel and username.
The webhook url can be found in the 'Incoming Webhooks' settings for our
Slack server and should be kept private at all times.

The channel can be changed to allow for posting to your private messages
for testing purposes. With the username being there for a better
distinction between staging etc so it could appear as 'Alces Flight
Center [Staging]' for

All of these can be set in your .env file which has an updated example
for the Slack variables attached in this commit.