Run migrations in pre rather than post deploy hook [skip ci]
According to @benarmston at
https://alces.slack.com/archives/C72GT476Y/p1528188195000103, the Dokku
`postdeploy` hook runs after the app container is started and so using
it could potentially still lead to us having the same issue as before
(ActiveRecord having outdated knowledge of the database, leading to odd
errors).

I don't know if this would actually cause us an issue or not, but if
using `predeploy` works fine then let's remove the possibility of this
issue occurring by using it.