Dynamically define host to use when generate URLs
Rather than hard-coding this to the production site host, use the same
host as configured for the environment for mailers; this means URLs in
e.g. Slack messages should be correct for the environment which
generated them.

Doing this also seems to remove the need for an earlier hack I added, or
at least the tests continue passing without the hack after this change.