Explicitly set the slack username in development
This removes the need to assign one of the Slack environment variables
as there is no real need for people to have a unique name for the
notifications. The only thing that is important is that it states it is
from development so it can be distinguished from other notifications.