Verify authorization occurs in all controllers
I have decided that, for now at least, we will require authorization for
all non-read-only controller actions, since these are the actions which
certain Users may be forbidden from accessing.

Therefore this commit moves verifying that authorization occurs for
these actions to the `ApplicationController`, with a couple of special
cases in some other individual controllers (see `SsoSessionsController`
and `CasesController`). We can also then remove the `verify_authorized`
calls from other controllers, as well as the now unneeded authorization
as per the previous paragraph.

Later we may look into moving our read permissions checks (see
`ApplicationRecord#check_read_permissions`) from `ApplicationRecord` to
policies and/or controllers, but this seems like it may be somewhat
involved and is fairly orthogonal to why we are originally introducing
authorization via Pundit (so we can consistently handle access to
different actions for the various types of Users we now have), and I'm
not sure if we want to do this in any case.