Create role scopes and use in SiteDecorator
For each User role, create a corresponding scope. This means we don't
need to create a method to access just the associated Users with a
particular role in Site, or wherever else we might want to do this in
future. We can also then use this scope in the SiteDecorator, which also
has the advantage that we now cache the secondary contacts, to avoid
making several queries for the same data (which we did previously).