Switch existing inline auth checking to use Pundit
This commit changes various simple cases where we previously performed
inline authorisation checking based on User methods, in both views or
decorators, to now check authorisation using Pundit and the recently
implemented Pundit policies.

This should be functionally equivalent, but means we are no longer
duplicating this logic between the policies and views/decorators; this
means if we later want to change when a User is authorised to perform an
action this can be done in one easily found place rather than needing to
search through views.