Set up rspec-retry, and use for JS tests
Sometimes tests which run JS (which are currently just in
`spec/features/case/new_spec.rb`) fail intermittently, I think just due
to there being a lot of inherent complexity to this sort of test rather
than anything actually wrong/fixable with how these are written. This is
particularly annoying when this happens in CI as the whole build must
then be restarted to check the suite can pass, possibly multiple times
if you are particularly unlucky.

This commit sets up the `rspec-retry` Gem and uses it when running these
type of tests, which should hopefully reduce this issue.

We do still have some other tests which very intermittently fail, but
just always retrying these, rather than working out and fixing the
underlying issues, seems like it would be a hack rather than the correct
fix, so this has not been done for now - we could do this later if any
of these started annoying us at all though.

Related card:
https://trello.com/c/YmGIOtap/273-fix-intermittent-ci-failures.