Show buttons based on available transitions
Rather than based on the current state of the Case. This is functionally
equivalent for now, but means we show the button to initiate a
transition if and only if that transition is possible rather than
relying on the state and our current knowledge of the state machine,
which could become outdated if the Case state machine changes.