Hide new Case link in dropdown from viewers
Viewers can't create Cases, so no need to show this link.

Also needed to refactor TabsBuilder to also be passed the User to build
the tabs for to do this; previously it was only able to access the User
via the decorated scope, and the User available within this was stubbed
by the tests.