Use `button_tag` to create Case escalate button
We need to conditionally disable this button depending on whether the
User can actually escalate the Case, and using Rails' `button_tag`
helper rather than creating the HTML directly should make doing this
slightly simpler and more consistent with other places in the app.

Note that we don't want to use `button_to` here as we don't actually
want the button to submit anywhere, rather we want it just to show a
modal.