Display MOTD tool on page load
This works by having the `init` function return a list of `Cmd Msg`s to
display the desired category and issue.

Whist this works, it would be better to have `decodeInitialModel` set
the desired category and issue from `flags`.  Doing that would avoid a
couple of trips through the elm runtime.  Unfortunately, how to do that
is not immediately obvious to me.