Extract shared examples for viewer-disabled button
We want to similarly disable various buttons (or button links) when the
actions they correspond with are disabled for viewers but enabled for
contacts, therefore extract shared examples to ensure these are tested
and implemented (at least from an external perspective) in similar ways.

Note that to have these function similarly also needed to add the
`disabled` class in the case of the escalate Case button; although not
necessary for a button (as opposed to a button link), and has no effect
on the appearance, this makes the handling in each case consistent. This
should later make extracting shared code to handle the implementation in
the different cases more straightforward.