Change confusing nesting of styles
I put these inside the `body {...}` group as this was a way to get them
to override the styles for the same selectors in Bootstrap, by giving
them a higher specificity. But this now seems like a slight hack since
it's not obvious why these need to be here and that moving them would
break the styling by making them not apply.

Therefore move them outside and just use `!important`, which though not
usually a good idea seems more correct since we do want them to always
override other styles.