Extract class for modifying options based on policy
In several cases, and with more to come soon, we want to disable but
still show buttons/button links to Users who do not have the appropriate
authorisation to perform that action. Therefore this commit extracts a
helper class to perform this modification to the options to be used when
creating these buttons/button links etc.

Note that this is done by modifying the `html_options` rather than
creating a new helper method to create a button of this kind, as there
are various Rails helper methods we may want to use to create these
buttons/button links, all of which accept the same options, so this
seems a better place to do the modification (e.g. we use `link_to`,
`button_to`, `button_tag` etc.).