Add slight delay to showing tooltips
The intention here is that tooltips should only show if the user
intentionally hovers over the element for a brief amount of time.
Avoiding tooltips displaying when simply scrolling through a page and
mousing over tooltip elements for a split second.

However the exact timings here may need some adjustments to get the
timing to feel natural.