Add coverage for couple of NoteDecorator methods
In preparation for changing these for viewers.

Also needed to change the code slightly to function the same but be more
testable, by: accessing `current_user` via Draper's `h` access to
helpers (accessing the method on the decorator directly worked in views
as `current_user` is made available directly there, but won't work by
default outside views); and also squish text, so it displays the same
but uses consistent spacing, and so is more matchable.