Fix tests failing when run on Friday after 5 o_O
Fixes a similar issue to
https://github.com/alces-software/alces-flight-center/commit/245be6ba554c68738d4a1a064106bfcc1fe0c12e,
but this should hopefully be a more robust fix. The underlying issue is
that these tests had a dependency on the real time, and so could pass or
fail depending on what the real time is when they are run. Ensuring we
always use the same fake time should at least make them deterministic.