Allow missing details to cause failures
Reverted this change to allow these situations to 'fail fast'. I have
also added a test for the CR event card(s) which would have picked up on
this issue.

However this does not resolve the main problem at hand. If a new event
is added that omits `details` or anything else that is vital then they
will fail silently if nobody adds a test for them initially. This means
that erroneous event cards could still slip through.