Add a cluster checks widget to cluster overview
This will contain information on how many checks were successful for the
last set of checks, with the date of the checks displayed on the line
below. As well as allowing easy navigation to view that check, or
complete it if the user is an admin and the checks for that day have not
been done.