FlightDirectoryConfig model et al
This model stores the configuration necessary for
flight-terminal-service to be able to SSH into the Flight Directory
appliance for the current user's site.

The config is only available to site contacts.  Other users requesting
the config will not be provided with it and flight-terminal-service will
not connect to the Directory service.