Add capacity for encrypting database values
Flight Center stores a public RSA key which will be used to encrypt the
private SSH keys for the terminal service.  The terminal service will
have the corresponding private RSA key which it can use to decrypt the
SSH key.

There are a couple of gotchas to be aware of when generating the keypair:

 1. We want to make sure that the public and private key are never
    stored in the same location, including any log files.

 2. Using asymmetric encryption for the SSH key requires that we use an
    unusually large RSA key pair.

A rake task has been added which will create a suitably large RSA key.
Steps have been taken to prevent accidentally running the rake task in a
way that the keypair would be stored in log files together.

There is another gotcha with regard to padding.  The default paddings
differ between our Ruby and Nodejs OpenSSL bindings, being explicit in
both locations avoids any surprises.