Make multiple filters work with each other.
If we use Popovers as previous, then each part of the filter form only
exists in the DOM as long as the popover is open. Close the popover and
lose the filter (even if you've not changed it).

By using Bootstrap's dropdowns instead the DOM form remains constant,
existing filters are maintained, and all is well with the world.

One must remember to have an `onclick="event.stopPropagation()"` event
handler on anything interactive within the dropdown else clicking will
close the dropdown (not helpful).