Filter by case associations.
Hopefully intuitively, associations are `or`ed together, then `and`ed
with any other filters. This is exactly how the other filters work, but
it takes a bit more work to get there since we need to look through an
association.

We need to apply the other filters _after_ the association filtering is
done.