Dynamically change the text colour
To match the credit balance widget the text colour for the cluster
checks one is also now dynamic. With it displaying as red if no checks
have passed, orange if less than half have passed and green if half or
more have passed.

These values are subject to change and can be modified very easily so
don't consider these final thresholds.