Import a minimal set of Bootstrap styles.
Flight Center actually gets most of its Bootstrap styling via Flight
Chrome; we don't otherwise import Bootstrap for the SCSS processor. But
here, we need to let the processor access some Bootstrap definitions for
us to successfully extend them.

This is a minimal set of Bootstrap imports for the table definitions to
work.