Only show checks as green if all have passed
The threshold for check result classes has been changed slightly so that
on the dashboard the text is only green if and when all checks have
passed for that day. Below that the text is orange until all checks have
failed, at which point it will be red.

Again these values are subject to change, especially for the red text
which could potentially be made a bit higher.