Only add event listeners to the correct editor
configureMarkdownPreview was adding event listeners for each editor on
the page. This meant that when it was being called after every
showPreview and showWrite action it would unnecessarily add event
listeners to every editor again.

This commit ensures that event listeners are only being added to the
editor that has just been interacted with.