Ensure id is Integer before checking if positive
If the `CasesController#create` action is triggered via an Ajax request
from the new Case form then this worked fine, since the params are JSON
and this will already be an Integer. However if it was triggered via a
regular POST from a change support type button then the `service_id`
will be a String, and this method call would fail.