Fix tests and add coverage for injected Service
Adapt the existing tests and add new ones covering the injection of the
'Other' Service for the Case form. Prompted by this also handle the
situation where there are no Issues to be associated with the 'Other'
Service, which in turn fixes the test failures elsewhere in the suite
due to not handling this.