Set Sentry error context for rails-admin controllers
After extracting the RootController, for all controllers which will be
used by our app to inherit from (including those which come from other
Gems like rails-admin), the extra Sentry Raven context which we set in
the ApplicationController was no longer being applied to these other
controllers. This was then causing errors like
https://sentry.io/alces-software-ltd/alces-flight-center/issues/626456706
to not include this additional identifying context, such as the User who
triggered the exception.

The straightforward fix for this is to just extract setting this context
up a level to the RootController as well.