Use Html.Keyed to prevent Elm reusing selects
Key each `select` field in such a way that a new key is used each time
the parent field, and therefore the items within each `select`, change.
This prevents Elm from reusing different `select`s when this happens,
which in turn is useful as `selected` is only applied to a `select`
`option` on render, and so if a `select` is reused it can cause weird
behaviour with the selected option not corresponding with the selected
item in the model.