Don't default to cluster if case associations are empty.
Since we're now allowing cases to be created with no associations via
the "Other" pseudoservice/issue, it looks a bit jarring to then be
presented with "the entire cluster is affected!!!" the moment you
submit.

This removes the default association from `Case`s and adds placeholder
"None" text in places that would otherwise be empty as a result.