Validate Component selected, if required
We use a DrillDownSelectList for storing many items in the State,
however Components are the only items that we store in this way and that
are also displayed in an Unselected form at the point where it is
possible to submit the Case form (since the Components field is shown,
when required, when an Issue is Selected, and this also reveals all
fields within the Tier tabs including the submit button).

Therefore we need to validate that a Component has been selected when
required, to avoid the form being submitted without doing this; for
other items stored as a DrillDownSelectList we do not need to do this
since the form cannot be submitted without selecting them, however this
may change if the form structure later changes.