Create map function and use to fix mapSelected
`mapSelected` was previously always returning a Selected, which I don't
think was effecting anything, but this doesn't seem like what we'd
always want to happen (there's not particularly a reason why
transforming the selecting value itself in a DrillDownSelectList should
change it to Selected if it wasn't previously), and could effect things
later.

Therefore create a map function to map over the SelectList contained by
the DrillDownSelectList, and then use this to fix this. This is
currently only used internally within DrillDownSelectList, so is not
exposed for now.