Add comment explaining date format justification
After only wasting a small amount of time it was decided that, for now,
we should keep this the way it is as only the exact day of the results
is saved.

If at any point this requirement changes then we will need to adjust
this to work with proper UTC date and times.