Pre-select option when only one available
When constructing each DrillDownSelectList, if only one option is
available no need to require the user to select this, so initialize as
Selected, otherwise initialize as Unselected so they must choose.

As a consequence of doing this can also remove the special handling for
Clusters where we did a similar thing, since this will now already be
handled.

This resolves the first suggestion from
https://github.com/alces-software/alces-flight-center/pull/471#issuecomment-409989509.