Define datepickers for various use cases
The required datepickers for Checks, Credit Usage and Maintenance
Windows all have different needs, Therefore each one has a separately
defined datepicker with mostly unique intialisation.

Having them defined in this way allows for easier individual
customisation but is not 100% necessary. The changes required for all 3
can be handled by calling methods on the datepicker to adjust the
settings.

Also in this commit is a change to reflect the new id name for the
currently used datepicker for Checks.