Replace old date selection with the new datepicker
The old date selection method for depositing credits and selecting the
date for a maintenance window has been replaced with the js datepicker.

Making the switch for the Credit Usage page was relatively simple.
However it was a bit more involved for the Maintenance Request form and
so there are various changes to its controller etc to get this
functional.