Prevent potentially infinite Slack notifications
I was saving the variable containing an associated log for a cluster
check result as an instance variable. Because of this and the way in
which I was handling sending a Slack notification for the new log it
would send a notification for the same log for as many times as there
were cluster check results left to iterate through. Unless however a new
log was created, at which point the process would continue until the
end.

This commit prevents that behaviour and lets only one notification per
log go through.