Completely remove the concept of a component type.
We have no sensible way of getting this information from anywhere other
than tedious manual intervention, so we're dropping it for now. We might
revisit it later (it'd be nice to have different icons for different
types of components).

The cluster components dropdown is now grouped by, um, group rather than
by type. We could probably revisit the navigation in this area too.