Update readme
The purpose of this added step for setting up a dev environment is so
that any current and future environment variables are accounted for. For
example as of this commit the `TERMINAL_SERVICE_BASE_URL` is required
now that the Flight terminal service is enabled on the Demo site, with
more sites likely to follow in the future.