Show both `Other`s in the list of possible issues if we have a service.
Otherwise (no pun intended) if we have a no-service-Other case and add a
service association, then then no-service-Other is not shown in the
dropdown which makes it look like it's got a completely different issue.

I'm going to deal with the confusion of having two indistinguishable
Other issues in the next commit.