Initialse case contact for all open cases
The migration in this commit should assign a contact to all currently
open cases. This follows the same logic as default contact assignment
when creating a case.

Therefore if the case was created by an admin the contact is now the
primary contact for the site. Else it is assigned to the user that
created the case.