Tweak how case reassignment works.
We don't want to do it pre-validation: because if validation fails, then
we'll have reassigned all the cases away from a user that is still a
contact. So we'll do it post-validation and just before the change to
`role` is saved.

However, this means that we need to ignore the case where the role has
just changed to be a viewer when we validate (and maintain the validity
of the user record with the before_save hook).